@@ -1,13 +1,33 @@
-import java.io.File;
-import java.util.Arrays;
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+
+import java.io.*;
+import java.net.ServerSocket;
+import java.net.Socket;
 
 public class Main {
     public static void main(String[] args) throws Exception {
+        GsonBuilder builderGson = new GsonBuilder();
+        Gson gson = builderGson.setPrettyPrinting().create();
         BooleanSearchEngine engine = new BooleanSearchEngine(new File("pdfs"));
-        System.out.println(engine.search("бизнес"));
+        //System.out.println(engine.search("бизнес"));
 
-        // здесь создайте сервер, который отвечал бы на нужные запросы
-        // слушать он должен порт 8989
-        // отвечать на запросы /{word} -> возвращённое значение метода search(word) в JSON-формате
+        System.out.println("Starting server at 8989 ...");
+        try (ServerSocket serverSocket = new ServerSocket(8989)) {
+            while (true) {
+                try (
+                        Socket socket = serverSocket.accept();
+                        BufferedReader in = new BufferedReader(new InputStreamReader(socket.getInputStream()));
+                        PrintWriter out = new PrintWriter(socket.getOutputStream())
+                ) {
+                    //out.println(gson.toJson(engine.search(in.readLine())));
+                    out.println(gson.toJson(engine.searchRequest(in.readLine(), "stop-ru.txt")));
+                }
+            }
+        } catch (IOException e) {
+            System.out.println("Не могу стартовать сервер");
+            e.printStackTrace();
+        }
+
     }
 }
