@@ -3,5 +3,36 @@ public class PageEntry implements Comparable<PageEntry> {
     private final int page;
     private final int count;
 
-    // ???
+    public PageEntry(String pdfName, int page, int count) {
+        this.pdfName = pdfName;
+        this.page = page;
+        this.count = count;
+    }
+
+    public PageEntry countSum(PageEntry pageEntry) {
+        int countSum = this.count + pageEntry.count;
+        return new PageEntry(this.pdfName, this.page, countSum);
+    }
+
+    @Override
+    public int compareTo(PageEntry o) {
+        return Integer.compare(o.count, this.count);
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj) return true;
+        if (obj == null || getClass() != obj.getClass()) return false;
+        PageEntry pageEntry = (PageEntry) obj;
+        return pdfName.equals(pageEntry.pdfName) && page == pageEntry.page;
+    }
+
+    @Override
+    public String toString() {
+        return "PageEntry{" +
+                "pdfName='" + pdfName + '\'' +
+                ", page=" + page +
+                ", count=" + count +
+                '}';
+    }
 }
